# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Personal Finance API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/v1/users/wallets/{wallet_id}/transaction": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Crear una transacción en la cartera del usuario",
                "operationId": "5a3c8eb223f298539ae36695f46e76cd",
                "parameters": [
                    {
                        "name": "wallet_id",
                        "in": "path",
                        "description": "ID de la cartera",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "ingreso"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 100.5
                                    },
                                    "category": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Pago de salario"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transacción creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transacción creada exitosamente"
                                        },
                                        "transaction": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "wallet_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "transaction_type": {
                                                    "type": "string",
                                                    "example": "ingreso"
                                                },
                                                "category_id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "example": 100.5
                                                },
                                                "balance_after_transaction": {
                                                    "type": "number",
                                                    "example": 1100.5
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-11-06T19:10:44Z"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Pago de salario"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos de transacción no válidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Monto no válido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error inesperado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error inesperado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/wallets/{wallet_id}/transactions": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Obtener todas las transacciones de una cartera específica",
                "operationId": "5914a5bfe0e37ffdf14f9f30ec90ba2a",
                "parameters": [
                    {
                        "name": "wallet_id",
                        "in": "path",
                        "description": "ID de la cartera",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de transacciones de la cartera",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "wallet": {
                                            "type": "string",
                                            "example": "Bancaria"
                                        },
                                        "transactions": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "wallet_name": {
                                                        "type": "string",
                                                        "example": "Bancaria"
                                                    },
                                                    "type_transaccion": {
                                                        "type": "string",
                                                        "example": "ingreso"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "Salario"
                                                    },
                                                    "ammount": {
                                                        "type": "string",
                                                        "example": "$100.50"
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-11-06T19:10:44Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cartera o transacciones no encontradas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No hay transacciones en esta cartera"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error inesperado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error inesperado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Se requiere un token de acceso Bearer JWT",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}